Windows平台的窗口的RGN设置流程优化：圆角使用与Skia一致的RGN，从而避免窗口的圆角出现黑点的问题（出现条件：非分层窗口设置圆角，根容器设置圆角，两个圆角相同）。 代码重构，RGN相关的代码封装在一个类中，同时当使用SDL的时候，在Windows平台仍然生效。
