+++ duilib/RenderSkia/Render_Skia_Windows.cpp
@@ -4,14 +4,13 @@
 
 #include "SkRasterWindowContext_Windows.h"
 #include "SkGLWindowContext_Windows.h"
+#include "WindowRgn_Windows.h"
 
 #include "SkiaHeaderBegin.h"
 
 #include "include/core/SkCanvas.h"
 #include "include/core/SkSurface.h"
 #include "include/core/SkRegion.h"
-#include "include/core/SkPath.h"
-#include "include/core/SkPathMeasure.h"
 
 #include "SkiaHeaderEnd.h"
 
@@ -309,99 +308,19 @@ void Render_Skia_Windows::DeleteDC()
     }
 }
 
-// 从Skia的圆角路径生成HRGN，保证与drawRoundRect完全一致
-static HRGN CreateHRGNFromSkiaRoundRect(const SkRect& skRect, SkScalar rx, SkScalar ry)
-{
-    // 创建Skia圆角路径（确保路径闭合）
-    SkPath path;
-    path.addRoundRect(skRect, rx, ry);
-    path.close(); // 显式闭合路径，避免端点缺失
-
-    // 使用SkPathMeasure离散化路径为密集点集
-    SkPathMeasure pathMeasure(path, true); // true表示闭合路径
-    std::vector<POINT> points;
-
-    // 路径总长度
-    SkScalar pathLength = pathMeasure.getLength();
-    if (pathLength <= 0) {
-        return nullptr;
-    }
-
-    // 离散化步长（越小越精细，建议1.0f，兼顾精度和性能）
-    const SkScalar step = 1.0f;
-    SkScalar distance = 0.0f;
-
-    // 遍历路径，按步长提取点
-    while (distance <= pathLength) {
-        SkPoint pos;
-        SkVector tan;
-        // 获取指定距离处的点坐标和切线
-        if (pathMeasure.getPosTan(distance, &pos, &tan)) {
-            // 转换为Windows POINT（浮点转整数，四舍五入更精准）
-            POINT pt = {
-                static_cast<LONG>(std::round(pos.x())),
-                static_cast<LONG>(std::round(pos.y()))
-            };
-            // 避免重复添加相同点
-            if (points.empty() || !(points.back().x == pt.x && points.back().y == pt.y)) {
-                points.push_back(pt);
-            }
-        }
-        distance += step;
-    }
-
-    // 移除最后一个重复的起点（路径闭合时会重复）
-    if ((points.size() >= 2) &&
-        (points.front().x == points.back().x) &&
-        (points.front().y == points.back().y)) {
-        points.pop_back();
-    }
-
-    // 验证点集有效性（至少3个点才能形成多边形）
-    if (points.size() < 3) {
-        return nullptr;
-    }
-
-    // 创建多边形HRGN（ALTERNATE填充模式适配圆角）
-    HRGN hRgn = ::CreatePolygonRgn(points.data(), static_cast<int>(points.size()), ALTERNATE);
-    return hRgn;
-}
-
 bool Render_Skia_Windows::SetWindowRoundRectRgn(const UiRect& rcWnd, float rx, float ry, bool bRedraw)
 {
-    ASSERT(::IsWindow(m_hWnd));
-    if (!::IsWindow(m_hWnd)) {
-        return false;
-    }
-    SkRect skRect;
-    skRect.fLeft = 0;
-    skRect.fTop = 0;
-    skRect.fRight = skRect.fLeft + rcWnd.Width();
-    skRect.fBottom = skRect.fTop + rcWnd.Height();
-
-    HRGN hRgn = CreateHRGNFromSkiaRoundRect(skRect, rx, ry);
-    int nRet = ::SetWindowRgn(m_hWnd, hRgn, bRedraw ? TRUE : FALSE);
-    return nRet != 0;
+    return WindowRgn::SetWindowRoundRectRgn(m_hWnd, rcWnd, rx, ry, bRedraw);
 }
 
 bool Render_Skia_Windows::SetWindowRectRgn(const UiRect& rcWnd, bool bRedraw)
 {
-    ASSERT(::IsWindow(m_hWnd));
-    if (!::IsWindow(m_hWnd)) {
-        return false;
-    }
-    HRGN hRgn = ::CreateRectRgn(rcWnd.left, rcWnd.top, rcWnd.right, rcWnd.bottom);
-    int nRet = ::SetWindowRgn(m_hWnd, hRgn, bRedraw ? TRUE : FALSE);
-    return nRet != 0;
+    return WindowRgn::SetWindowRectRgn(m_hWnd, rcWnd, bRedraw);
 }
 
 void Render_Skia_Windows::ClearWindowRgn(bool bRedraw)
 {
-    ASSERT(::IsWindow(m_hWnd));
-    if (!::IsWindow(m_hWnd)) {
-        return;
-    }
-    ::SetWindowRgn(m_hWnd, nullptr, bRedraw ? TRUE : FALSE);
+    return WindowRgn::ClearWindowRgn(m_hWnd, bRedraw);
 }
 
 } // namespace ui
